requirements.txt
loguru
gspread
gspread_asyncio
aiohttp
pandas
google-auth
pytz
python-telegram-bot
python-dotenv
asyncio